Well, memosono was adding up memory usage by each click on a tale. This is what a normal human would do as well when playing the memory game;) But the advantag of computers is...just kidding.
The problem was the usage of pixbuf. A new pixbuf was created by each click and the memory did not gets deleted after the object was out of scope. Fixes were the addition of those two lines:

del ppixbuf_i
gc.collect()

Deleting the pixbuf and calling the garbage collector manually.

Bug with the best story.