Replace the combo used to change the font by a button with a palette
This is needed to use with touch.
As the combo was removed, is not needed hide the labels with the screen
rotated anymore. The RoundBox containing the buttons in the CardEditors
was replaced by a HBox due to size concerns.
The list of fonts is read from the configuration files.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>