Reorbanize main screen and scoreboard
Now, the scoreboard is always below the table game.
That limit the space to show users, but anyway we are allowing play
only to 2 players, due to problems with the collaboration.
This reorganization try to put the focus in the game space.
More work is neede to center the board.

This commit also replace the use of SvgLabel for a standard sugar Icon,
then, remove two files and simplify a lot the code on the scoreboard.