Remove theme.py
The package had only common definitions.
In some cases, the variables was not used at all,
other were used in only one place. Other could be replaced by constants
on sugar3.style

The variables were moved to the package when are used.
Now the constanst are idependient on the resolution
as the constants defined on style.