Fix robot face display
Now, nstead of use face.py widget, draw the speak icon.
This solve the complicate substitution used to show the robot face.
Before the port to gtk3, the face eyes followed the cursor,
but that is not working after the port. And with the flio animation
is more difficult implement that again, and do not add too much value.

The cement method do nothing, and can be removed.